Mutapox hediff set upd
Fixed a bug due to which mutapox was not immunized by genes.

Also the code was optimized a little. The event should no longer make frequent checks. And its impact on performance should be significantly reduced, almost to zero.

The frequency of the event itself was increased, but it should actually become a little less frequent, since toxic resistance reduces the chance of getting mutapox.