Mutapox incident
Mutapox trigger has been replaced. It will now trigger as a disease incident, when the map is heavily polluted.

The chance starts to increase from 25%, is quite high at 50% and is almost guaranteed at 80+%.

The chance of choosing a disease can be changed in Patches_HumanHediffSet. The chance of disease in BMT_Incidents